feat(nixvim): add nvim-lightbulb and ui-select plugins
This commit introduces two new plugins to the nixvim configuration.
Firstly, the nvim-lightbulb plugin is added with autocmd enabled.
Secondly, the ui-select plugin for telescope is enabled.

The keybinding for code actions is also updated. The previous
binding that used actions-preview is replaced with a direct
code_action command.

Lastly, some redundant code related to code actions is removed.