refactor: rearrange key bindings in haskell-tools-nvim.nix
This commit rearranges the key bindings in the haskell-tools-nvim.nix
file. The changes include swapping the bindings for Hoogle search and
code snippet evaluation, and adding new bindings for opening the package
cabal file and the project file. This rearrangement aims to improve the
usability and efficiency of the key bindings.