feat(nixvim): add neotest plugin and keymaps
This commit introduces the neotest plugin to the nixvim configuration.
It enables the plugin and sets the log level to "trace". It also enables
the Haskell adapter for the plugin.

Several keymaps are added for running tests, opening the test output
window, toggling the test summary, and attaching to a running test.