feat(mini.nix): add split window functionality to MiniFiles
This commit introduces a new feature in the MiniFiles plugin. It allows
users to split the window either horizontally or vertically. The keys
'gs' and 'gv' have been mapped to perform the belowright horizontal and
belowright vertical splits respectively. This feature enhances the
usability of the MiniFiles plugin by providing more flexibility in
window management.