feat(nixvim): add conform-nvim and remove format keymap from lsp.nix
This commit introduces conform-nvim to the nixvim configuration. It
provides a comprehensive set of formatters for various file types and
sets up a keymap for formatting files or ranges. The previous format
keymap in lsp.nix has been removed as it's now handled by conform-nvim.

Additionally, the incremental_selection feature has been enabled in
the default.nix file.